backend/requirements.txt
flask
flask-cors
python-dotenv
werkzeug
pdfplumber
phonenumbers
spacy
python-docx
supabase